fixed sircular dep
