+++ src/echo-core/types/components.ts
@@ -1,5 +1,5 @@
 import { ComponentType } from 'react';
-import { BaseAppComponentProps } from './api';
+import { RouteComponentProps } from 'react-router-dom';
 import { Without } from './common';
 
 /**
@@ -8,3 +8,18 @@ import { Without } from './common';
 export type AnyComponent<T> = ComponentType<T>;
 
 export type WrappedComponent<TProps> = ComponentType<Without<TProps, keyof BaseAppComponentProps>>;
+
+/**
+ * The props that every registered App component obtains.
+ */
+export interface RouteBaseProps<UrlParams = unknown, UrlState = unknown>
+    extends RouteComponentProps<UrlParams, {}, UrlState>,
+        BaseAppComponentProps {}
+
+/**
+ * The Base Components Props obtained by avery app.
+ */
+export interface BaseAppComponentProps {
+    name?: string;
+    // this Interface needs better definement.
+}
